Extract topological sort into a separate util
